requirements.txt
streamlit
pandas
python-dotenv
faiss-cpu
sentence-transformers
google-genai
dotenv